First commit
Proof of concept to scrape the page for Sensei's Divining Top.
Should be made into a function to take the URL for a card and
return the copus of information about the sellers of the card,
subject to supplied conditions (which can be submitted on the page
itself, before scraping).